requirements.txt
surprise==0.1
pandas==1.1.1
numpy==1.18.1
Flask==1.1.1
pandas-gbq==0.14.0
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-cloud-bigquery==2.1.0
tqdm==4.41.1
gunicorn
